Unify FilterPresubmit logic and move to pjutil.
