+++ prow/plugins/trigger/generic-comment_test.go
@@ -18,7 +18,6 @@ package trigger
 
 import (
 	"fmt"
-	"k8s.io/test-infra/prow/pjutil"
 	"log"
 	"reflect"
 	"testing"
@@ -1494,237 +1493,6 @@ func TestDetermineSkippedPresubmits(t *testing.T) {
 		t.Run(testCase.name, func(t *testing.T) {
 			if actual, expected := determineSkippedPresubmits(testCase.toTrigger, testCase.toSkipSuperset, logrus.WithField("test-case", testCase.name)), testCase.expectedToSkip; !reflect.DeepEqual(actual, expected) {
 				t.Errorf("%s: incorrect skipped presubmits determined: %v", testCase.name, diff.ObjectReflectDiff(actual, expected))
-			}
-		})
-	}
-}
-
-type fakeChangesGetter struct {
-	shouldError bool
-}
-
-func (c *fakeChangesGetter) GetPullRequestChanges(org, repo string, number int) ([]github.PullRequestChange, error) {
-	if c.shouldError {
-		return nil, errors.New("error getting changes")
-	}
-	return nil, nil
-}
-
-func TestFilterPresubmits(t *testing.T) {
-	var testCases = []struct {
-		name                              string
-		filter                            pjutil.Filter
-		presubmits                        []config.Presubmit
-		changesError                      bool
-		expectedToTrigger, expectedToSkip []config.Presubmit
-		expectErr                         bool
-	}{
-		{
-			name: "nothing matches, nothing to run or skip",
-			filter: func(p config.Presubmit) (shouldRun bool, forcedToRun bool, defaultBehavior bool) {
-				return false, false, false
-			},
-			presubmits: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "ignored"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "ignored"},
-				Reporter: config.Reporter{Context: "second"},
-			}},
-			changesError:      false,
-			expectedToTrigger: nil,
-			expectedToSkip:    nil,
-			expectErr:         false,
-		},
-		{
-			name: "everything matches and is forced to run, nothing to skip",
-			filter: func(p config.Presubmit) (shouldRun bool, forcedToRun bool, defaultBehavior bool) {
-				return true, true, true
-			},
-			presubmits: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "second"},
-			}},
-			changesError: false,
-			expectedToTrigger: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "second"},
-			}},
-			expectedToSkip: nil,
-			expectErr:      false,
-		},
-		{
-			name: "error detecting if something should run, nothing to run or skip",
-			filter: func(p config.Presubmit) (shouldRun bool, forcedToRun bool, defaultBehavior bool) {
-				return true, false, false
-			},
-			presubmits: []config.Presubmit{{
-				JobBase:             config.JobBase{Name: "errors"},
-				Reporter:            config.Reporter{Context: "first"},
-				RegexpChangeMatcher: config.RegexpChangeMatcher{RunIfChanged: "oopsie"},
-			}, {
-				JobBase:  config.JobBase{Name: "ignored"},
-				Reporter: config.Reporter{Context: "second"},
-			}},
-			changesError:      true,
-			expectedToTrigger: nil,
-			expectedToSkip:    nil,
-			expectErr:         true,
-		},
-		{
-			name: "some things match and are forced to run, nothing to skip",
-			filter: func(p config.Presubmit) (shouldRun bool, forcedToRun bool, defaultBehavior bool) {
-				return p.Name == "should-trigger", true, true
-			},
-			presubmits: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "ignored"},
-				Reporter: config.Reporter{Context: "second"},
-			}},
-			changesError: false,
-			expectedToTrigger: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}},
-			expectedToSkip: nil,
-			expectErr:      false,
-		},
-		{
-			name: "everything matches and some things are forced to run, others should be skipped",
-			filter: func(p config.Presubmit) (shouldRun bool, forcedToRun bool, defaultBehavior bool) {
-				return true, p.Name == "should-trigger", p.Name == "should-trigger"
-			},
-			presubmits: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "second"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "third"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "fourth"},
-			}},
-			changesError: false,
-			expectedToTrigger: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "second"},
-			}},
-			expectedToSkip: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "third"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "fourth"},
-			}},
-			expectErr: false,
-		},
-		{
-			name: "everything matches and some things are forced to run, others should be skipped, overlapping contexts filtered out",
-			filter: func(p config.Presubmit) (shouldRun bool, forcedToRun bool, defaultBehavior bool) {
-				return true, p.Name == "should-trigger", p.Name == "should-trigger"
-			},
-			presubmits: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "second"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "fourth"},
-			}},
-			changesError: false,
-			expectedToTrigger: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "second"},
-			}},
-			expectedToSkip: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "fourth"},
-			}},
-			expectErr: false,
-		},
-		{
-			name: "everything matches and some things are forced to run, others should be skipped, overlapping contexts filtered out even if skipped comes first in the list",
-			filter: func(p config.Presubmit) (shouldRun bool, forcedToRun bool, defaultBehavior bool) {
-				return true, p.Name == "should-trigger", p.Name == "should-trigger"
-			},
-			presubmits: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "second"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "fourth"},
-			}},
-			changesError: false,
-			expectedToTrigger: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "first"},
-			}, {
-				JobBase:  config.JobBase{Name: "should-trigger"},
-				Reporter: config.Reporter{Context: "second"},
-			}},
-			expectedToSkip: []config.Presubmit{{
-				JobBase:  config.JobBase{Name: "should-skip"},
-				Reporter: config.Reporter{Context: "fourth"},
-			}},
-			expectErr: false,
-		},
-	}
-
-	pr := &github.PullRequest{
-		Base: github.PullRequestBranch{
-			Repo: github.Repo{
-				Owner: github.User{
-					Login: "org",
-				},
-				Name: "repo",
-			},
-			Ref: "foobar",
-		},
-		Number: 1,
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			actualToTrigger, actualToSkip, err := filterPresubmits(testCase.filter, &fakeChangesGetter{shouldError: testCase.changesError}, pr, testCase.presubmits, logrus.WithField("test-case", testCase.name))
-			if testCase.expectErr && err == nil {
-				t.Errorf("%s: expected an error filtering presubmits, but got none", testCase.name)
-			}
-			if !testCase.expectErr && err != nil {
-				t.Errorf("%s: expected no error filtering presubmits, but got one: %v", testCase.name, err)
-			}
-			if !reflect.DeepEqual(actualToTrigger, testCase.expectedToTrigger) {
-				t.Errorf("%s: incorrect set of presubmits to skip: %s", testCase.name, diff.ObjectReflectDiff(actualToTrigger, testCase.expectedToTrigger))
-			}
-			if !reflect.DeepEqual(actualToSkip, testCase.expectedToSkip) {
-				t.Errorf("%s: incorrect set of presubmits to skip: %s", testCase.name, diff.ObjectReflectDiff(actualToSkip, testCase.expectedToSkip))
 			}
 		})
 	}
